backend/requirements.txt
# FastAPI & Server
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
alembic==1.13.0

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0

# Claude AI
anthropic==0.7.7

# File handling
python-magic==0.4.27
PyPDF2==3.0.1
Pillow==10.1.0

# Excel generation
openpyxl==3.1.2

# Microsoft Graph (SharePoint)
msal==1.25.0
requests==2.31.0

# CORS
fastapi-cors==0.0.6

# Testing
pytest==7.4.3
httpx==0.25.2
